Move locally used code into the file that uses it
